refactor: move logging to shared
