+++ internal/provider/data_sources/public_cloud/instances/read.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
-	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/logging"
+	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/shared/logging"
 )
 
 func (d *instancesDataSource) Read(
@@ -19,7 +19,7 @@ func (d *instancesDataSource) Read(
 
 	if err != nil {
 		resp.Diagnostics.AddError("Unable to read instances", err.Error())
-		logging.FacadeError(
+		logging.ServiceError(
 			ctx,
 			err.ErrorResponse,
 			&resp.Diagnostics,
