+++ internal/provider/resources/public_cloud/instance/create.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
-	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/logging"
 	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/resources/public_cloud/model"
+	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/shared/logging"
 )
 
 func (i *instanceResource) Create(
@@ -27,7 +27,7 @@ func (i *instanceResource) Create(
 	if err != nil {
 		resp.Diagnostics.AddError("Error creating Instance", err.Error())
 
-		logging.FacadeError(
+		logging.ServiceError(
 			ctx,
 			err.ErrorResponse,
 			&resp.Diagnostics,
