+++ internal/provider/resources/public_cloud/instance/delete.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
-	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/logging"
 	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/resources/public_cloud/model"
+	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/shared/logging"
 )
 
 func (i *instanceResource) Delete(
@@ -37,7 +37,7 @@ func (i *instanceResource) Delete(
 			),
 		)
 
-		logging.FacadeError(
+		logging.ServiceError(
 			ctx,
 			err.ErrorResponse,
 			&resp.Diagnostics,
